Fix created user.status

+++ src/main/java/com/ronnachate/inventory/shared/constant/GenericConstant.java
@@ -1,5 +1,5 @@
 package com.ronnachate.inventory.shared.constant;
 
 public class GenericConstant {
-    public static final int ActiveUserStatus = 1;
+    public static final long ActiveUserStatus = 1;
 }
